Clear current word when we find a blank tile
When we encounter a blank tile, it's important to clear the current
word, otherwise we'll just append to it on the next iteration we find
a tile. This leads to incorrect words getting pushed; there can be
spaces between tiles and those will effectively be removed and all the
tiles will be treated together as a word which is clearly wrong.

Signed-off-by: Rick Elrod <rick@elrod.me>